+++ src/main/java/org/openlmis/fulfillment/repository/custom/OrderRepositoryCustom.java
@@ -25,7 +25,7 @@ import java.util.UUID;
 
 public interface OrderRepositoryCustom {
 
-  Page<Order> searchOrders(Set<UUID> supplyingFacility, UUID requestingFacility, UUID program,
+  Page<Order> searchOrders(Set<UUID> supplyingFacility, Set<UUID> requestingFacility, UUID program,
                            UUID processingPeriod, Set<OrderStatus> statuses, Pageable pageable);
 
   List<UUID> getRequestingFacilities(List<UUID> supplyingFacilityIds);